takt: issue-327-refactor-cli-skills
